refactor(desktop): share the headless agent run (#319)
The scheduler and the trigger dispatcher are Go's second and third callers of
agent.RunAgent, and the dispatcher is about to become the port's. Rather than
write the run twice, native/agent_run.rs holds it once.

The reason to share is sharper than tidiness. The run must go through
claude::client::query, the one-shot, and never through Session: a Session sets
session_mode, and the reader then deliberately neither closes stdin nor stops at
the result event, so a headless run's event channel never closes and the drain
blocks past the answer until the timeout. That shipped once in the scheduler and
was invisible to CI, to byte-identical live parity and to fifty unit tests,
because none of them ran a task. One implementation means the dispatcher cannot
reintroduce it.

The scheduler keeps what is its own — resolveSystemPrompt's strictness and the
timeout derived from timeout_minutes — and hands the rest over. All five
end-to-end scheduled-run tests still pass unchanged, which is the check that
matters for a refactor of this shape.
